[fleet] move fleet-related classes to flat "fleet" package
