+++ resources/META-INF/plugin-java.xml
@@ -23,9 +23,9 @@
       <action id="grammar.Generate.ParserUtil" class="org.intellij.grammar.actions.BnfGenerateParserUtilAction"/>
       <action id="grammar.Generate.JFlexLexer" class="org.intellij.grammar.actions.BnfGenerateLexerAction"/>
       <separator/>
-      <action id="grammar.fleet.Generate" class="org.intellij.grammar.actions.GenerateFleetAction"/>
-      <action id="grammar.fleet.Generate.JFlexLexer" class="org.intellij.grammar.actions.BnfGenerateFleetLexerAction"/>
-      <action id="grammar.fleet.Run.JFlex" class="org.intellij.grammar.actions.BnfRunFleetJFlexAction"/>
+      <action id="grammar.fleet.Generate" class="org.intellij.grammar.fleet.GenerateFleetAction"/>
+      <action id="grammar.fleet.Generate.JFlexLexer" class="org.intellij.grammar.fleet.BnfGenerateFleetLexerAction"/>
+      <action id="grammar.fleet.Run.JFlex" class="org.intellij.grammar.fleet.BnfRunFleetJFlexAction"/>
       <add-to-group group-id="grammar.file.group" anchor="first"/>
     </group>
   </actions>
