+++ src/org/intellij/grammar/fleet/FleetConstants.java
@@ -2,7 +2,7 @@
  * Copyright 2011-2024 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license that can be found in the LICENSE file.
  */
 
-package org.intellij.grammar.generator.fleet;
+package org.intellij.grammar.fleet;
 
 public interface FleetConstants {
   String GPUB_CLASS = "fleet.com.intellij.lang.parser.GeneratedParserUtilBase";
